# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The body does not contain required fields
      summary: Create a user with login and password
      tags:
        - Users
    get:
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: The list of users have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
      summary: Get all users
      tags:
        - Users
  /user/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The user has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Get single user
      tags:
        - Users
    put:
      operationId: UsersController_updatePassword
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The user's password has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The id is invalid (not uuid)
        '403':
          description: The oldPassword is wrong
        '404':
          description: The id doesn't exist
      summary: Update user's password
      tags:
        - Users
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The user has been successfully removed
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Delete user
      tags:
        - Users
  /artist:
    post:
      operationId: ArtistsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateArtistDto'
      responses:
        '201':
          description: The artist has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '400':
          description: The body does not contain required fields
      summary: Create an artist
      tags:
        - Artists
    get:
      operationId: ArtistsController_findAll
      parameters: []
      responses:
        '200':
          description: The list of artists have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtistResponse'
      summary: Get all artists
      tags:
        - Artists
  /artist/{id}:
    get:
      operationId: ArtistsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The artist has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Get single artist
      tags:
        - Artists
    put:
      operationId: ArtistsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateArtistDto'
      responses:
        '200':
          description: The artist has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Update the artist
      tags:
        - Artists
    delete:
      operationId: ArtistsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The artist has been successfully removed
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Delete artist
      tags:
        - Artists
  /track:
    post:
      operationId: TracksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateTrackDto'
      responses:
        '201':
          description: The track has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
      summary: Create a track
      tags:
        - Tracks
    get:
      operationId: TracksController_findAll
      parameters: []
      responses:
        '200':
          description: The list of tracks have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackResponse'
      summary: Get all tracks
      tags:
        - Tracks
  /track/{id}:
    get:
      operationId: TracksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The track has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Get single track
      tags:
        - Tracks
    put:
      operationId: TracksController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateTrackDto'
      responses:
        '200':
          description: The track has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Update the track
      tags:
        - Tracks
    delete:
      operationId: TracksController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The track has been successfully removed
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Delete track
      tags:
        - Tracks
  /album:
    post:
      operationId: AlbumsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateAlbumDto'
      responses:
        '201':
          description: The album has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: The body does not contain required fields
      summary: Create an album
      tags:
        - Albums
    get:
      operationId: AlbumsController_findAll
      parameters: []
      responses:
        '200':
          description: The list of albums have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlbumResponse'
      summary: Get all albums
      tags:
        - Albums
  /album/{id}:
    get:
      operationId: AlbumsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: The album has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Get single album
      tags:
        - Albums
    put:
      operationId: AlbumsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUpdateAlbumDto'
      responses:
        '200':
          description: The album has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponse'
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Update the album
      tags:
        - Albums
    delete:
      operationId: AlbumsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The album has been successfully removed
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The id doesn't exist
      summary: Delete album
      tags:
        - Albums
  /favs:
    get:
      operationId: FavoritesController_getFavorites
      parameters: []
      responses:
        '200':
          description: The list of favorites have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetFavoritesResponse'
      summary: Get all favorites
      tags:
        - Favorites
  /favs/track/{id}:
    post:
      operationId: FavoritesController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The track has been successfully added to favorites
          content:
            application/json:
              schema:
                type: string
        '400':
          description: The id is invalid (not uuid)
        '422':
          description: The id doesn't exist
      summary: Add the track to favorites
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The track has been successfully removed from favorites
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The track is not favorite
      summary: Delete track from favorites
      tags:
        - Favorites
  /favs/album/{id}:
    post:
      operationId: FavoritesController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The album has been successfully added to favorites
          content:
            application/json:
              schema:
                type: string
        '400':
          description: The id is invalid (not uuid)
        '422':
          description: The id doesn't exist
      summary: Add the album to favorites
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The album has been successfully removed from favorites
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The album is not favorite
      summary: Delete album from favorites
      tags:
        - Favorites
  /favs/artist/{id}:
    post:
      operationId: FavoritesController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: The artist has been successfully added to favorites
          content:
            application/json:
              schema:
                type: string
        '400':
          description: The id is invalid (not uuid)
        '422':
          description: The id doesn't exist
      summary: Add the artist to favorites
      tags:
        - Favorites
    delete:
      operationId: FavoritesController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: The artist has been successfully removed from favorites
        '400':
          description: The id is invalid (not uuid)
        '404':
          description: The artist is not favorite
      summary: Delete artist from favorites
      tags:
        - Favorites
info:
  title: REST Service
  description: Home Library Service
  version: '1.0'
  contact: {}
tags: []
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          minLength: 4
          example: John Dory
        password:
          type: string
          minLength: 4
          example: Qwerty123
      required:
        - login
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        version:
          type: number
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: Must match your old password
        newPassword:
          type: string
          minLength: 4
      required:
        - oldPassword
        - newPassword
    CreateUpdateArtistDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Mario
        grammy:
          type: boolean
      required:
        - name
        - grammy
    ArtistResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        grammy:
          type: boolean
      required:
        - id
        - name
        - grammy
    CreateUpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: I wanna be yours
        duration:
          type: number
          example: 360
        artistId:
          type: string
          example: null
        albumId:
          type: string
          example: null
      required:
        - name
        - duration
        - artistId
        - albumId
    TrackResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artistId:
          type: string
        albumId:
          type: string
        duration:
          type: number
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateUpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Dancing Shadows
        year:
          type: number
          example: 1996
        artistId:
          type: string
          example: null
      required:
        - name
        - year
        - artistId
    AlbumResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
      required:
        - id
        - name
        - year
        - artistId
    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        grammy:
          type: boolean
      required:
        - id
        - name
        - grammy
    Album:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: number
        artistId:
          type: string
      required:
        - id
        - name
        - year
        - artistId
    Track:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        artistId:
          type: string
        albumId:
          type: string
        duration:
          type: number
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    GetFavoritesResponse:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
      required:
        - artists
        - albums
        - tracks